requirements.txt
streamlit~=1.31.1
pandas~=2.2.1
streamlit_lottie
mysql-connector-python
requests~=2.31.0
pip~=23.2.1
attrs~=23.2.0
pillow~=10.2.0
tornado~=6.4
Jinja2~=3.1.3